fix ecds
