+++ pilot/pkg/model/telemetry.go
@@ -1012,142 +1012,6 @@ var waypointStatsConfig = protoconv.MessageToAny(&udpa.TypedStruct{
 // STOP. DO NOT UPDATE THIS WITHOUT UPDATING buildHTTPTelemetryFilter and buildTCPTelemetryFilter.
 const telemetryFilterHandled = 14
 
-// StatsProvider is the type of stats provider
-// Only prometheus and stackdriver are supported currently
-type StatsProvider string
-
-const (
-	StatsProviderPrometheus  StatsProvider = "prometheus"
-	StatsProviderStackdriver StatsProvider = "stackdriver"
-)
-
-type StatsConfig struct {
-	Provider         StatsProvider
-	NodeType         NodeType
-	ListenerClass    networking.ListenerClass
-	ListenerProtocol networking.ListenerProtocol
-}
-
-func (cfg StatsConfig) String() string {
-	return fmt.Sprintf("%s/%s/%v/%v", cfg.Provider, cfg.NodeType, cfg.ListenerClass, cfg.ListenerProtocol)
-}
-
-func StatsECDSResourceName(cfg StatsConfig) string {
-	return fmt.Sprintf("istio.io/telemetry/stats/%s", cfg.String())
-}
-
-func buildHTTPTypedExtensionConfig(class networking.ListenerClass, metricsCfg []telemetryFilterConfig) []*core.TypedExtensionConfig {
-	res := make([]*core.TypedExtensionConfig, 0, telemetryFilterHandled)
-	for _, cfg := range metricsCfg {
-		switch cfg.Provider.GetProvider().(type) {
-		case *meshconfig.MeshConfig_ExtensionProvider_Prometheus:
-			if cfg.NodeType == Waypoint {
-				res = append(res, &core.TypedExtensionConfig{
-					Name: StatsECDSResourceName(StatsConfig{
-						Provider:         StatsProviderPrometheus,
-						NodeType:         cfg.NodeType,
-						ListenerClass:    class,
-						ListenerProtocol: networking.ListenerProtocolHTTP,
-					}),
-					TypedConfig: protoconv.MessageToAny(waypointStatsConfig),
-				})
-			} else {
-				if statsCfg := generateStatsConfig(class, cfg); statsCfg != nil {
-					res = append(res, &core.TypedExtensionConfig{
-						Name: StatsECDSResourceName(StatsConfig{
-							Provider:         StatsProviderPrometheus,
-							NodeType:         cfg.NodeType,
-							ListenerClass:    class,
-							ListenerProtocol: networking.ListenerProtocolHTTP,
-						}),
-						TypedConfig: protoconv.MessageToAny(statsCfg),
-					})
-				}
-			}
-		case *meshconfig.MeshConfig_ExtensionProvider_Stackdriver:
-			sdCfg := generateSDConfig(class, cfg)
-			vmConfig := ConstructVMConfig("envoy.wasm.null.stackdriver")
-			vmConfig.VmConfig.VmId = stackdriverVMID(class)
-
-			wasmConfig := &wasm.PluginConfig{
-				RootId:        vmConfig.VmConfig.VmId,
-				Vm:            vmConfig,
-				Configuration: sdCfg,
-			}
-
-			res = append(res, &core.TypedExtensionConfig{
-				Name: StatsECDSResourceName(StatsConfig{
-					Provider:         StatsProviderStackdriver,
-					NodeType:         cfg.NodeType,
-					ListenerClass:    class,
-					ListenerProtocol: networking.ListenerProtocolHTTP,
-				}),
-				TypedConfig: protoconv.MessageToAny(wasmConfig),
-			})
-		default:
-			// Only prometheus and SD supported currently
-			continue
-		}
-	}
-	return res
-}
-
-func buildTCPTypedExtensionConfig(class networking.ListenerClass, metricsCfg []telemetryFilterConfig) []*core.TypedExtensionConfig {
-	res := make([]*core.TypedExtensionConfig, 0, telemetryFilterHandled)
-	for _, cfg := range metricsCfg {
-		switch cfg.Provider.GetProvider().(type) {
-		case *meshconfig.MeshConfig_ExtensionProvider_Prometheus:
-			if cfg.NodeType == Waypoint {
-				res = append(res, &core.TypedExtensionConfig{
-					Name: StatsECDSResourceName(StatsConfig{
-						Provider:         StatsProviderPrometheus,
-						NodeType:         cfg.NodeType,
-						ListenerClass:    class,
-						ListenerProtocol: networking.ListenerProtocolTCP,
-					}),
-					TypedConfig: protoconv.MessageToAny(waypointStatsConfig),
-				})
-			} else {
-				if statsCfg := generateStatsConfig(class, cfg); statsCfg != nil {
-					res = append(res, &core.TypedExtensionConfig{
-						Name: StatsECDSResourceName(StatsConfig{
-							Provider:         StatsProviderPrometheus,
-							NodeType:         cfg.NodeType,
-							ListenerClass:    class,
-							ListenerProtocol: networking.ListenerProtocolTCP,
-						}),
-						TypedConfig: protoconv.MessageToAny(statsCfg),
-					})
-				}
-			}
-		case *meshconfig.MeshConfig_ExtensionProvider_Stackdriver:
-			sdCfg := generateSDConfig(class, cfg)
-			vmConfig := ConstructVMConfig("envoy.wasm.null.stackdriver")
-			vmConfig.VmConfig.VmId = stackdriverVMID(class)
-
-			wasmConfig := &wasm.PluginConfig{
-				RootId:        vmConfig.VmConfig.VmId,
-				Vm:            vmConfig,
-				Configuration: sdCfg,
-			}
-
-			res = append(res, &core.TypedExtensionConfig{
-				Name: StatsECDSResourceName(StatsConfig{
-					Provider:         StatsProviderStackdriver,
-					NodeType:         cfg.NodeType,
-					ListenerClass:    class,
-					ListenerProtocol: networking.ListenerProtocolHTTP,
-				}),
-				TypedConfig: protoconv.MessageToAny(wasmConfig),
-			})
-		default:
-			// Only prometheus and SD supported currently
-			continue
-		}
-	}
-	return res
-}
-
 func buildHTTPTelemetryFilter(class networking.ListenerClass, metricsCfg []telemetryFilterConfig) []*hcm.HttpFilter {
 	res := make([]*hcm.HttpFilter, 0, len(metricsCfg))
 	for _, cfg := range metricsCfg {
